refactor: move form into entites
