+++ src/features/Callback/config.ts
@@ -1,5 +1,4 @@
-import { type TCallbackParams } from '@/features/Form';
-import { type TFieldConfig } from '@/features/Form/Field';
+import { type TCallbackParams, type TFieldConfig } from '@/entities/Form';
 
 import { validationTypes } from '@/shared/form';
