+++ src/features/EditBooking/EditBookingForm/ui.tsx
@@ -1,6 +1,6 @@
 import { FC, useState } from 'react';
 
-import { Form, type TBookingFormParams } from '@/features/Form';
+import { Form, type TBookingFormParams } from '@/entities/Form';
 
 import { StFlex, StTextBox } from '@/shared/styles/global';
 import { Button } from '@/shared/ui/Button';
