+++ src/features/MakeOrder/ui.tsx
@@ -1,6 +1,6 @@
 import { FC, useState } from 'react';
 
-import { Form, type TBookingFormParams } from '@/features/Form';
+import { Form, type TBookingFormParams } from '@/entities/Form';
 
 import { Button } from '@/shared/ui/Button';
