+++ src/features/RecoverPassword/config.ts
@@ -1,5 +1,4 @@
-import { type TRecoverPasswordParams } from '@/features/Form';
-import { type TFieldConfig } from '@/features/Form/Field';
+import { type TRecoverPasswordParams, type TFieldConfig } from '@/entities/Form';
 
 import { validationTypes } from '@/shared/form';
