+++ src/features/RecoverPassword/ui.tsx
@@ -1,6 +1,6 @@
 import { FC, useState } from 'react';
 
-import { Form, type TRecoverPasswordParams } from '@/features/Form';
+import { Form, type TRecoverPasswordParams } from '@/entities/Form';
 
 import { Button } from '@/shared/ui/Button';
