+++ src/widgets/user/NewPassword/config.ts
@@ -1,5 +1,4 @@
-import { TNewPassword } from '@/features/Form';
-import { type TFieldConfig } from '@/features/Form/Field';
+import { type TNewPassword, type TFieldConfig } from '@/entities/Form';
 
 import { validationTypes } from '@/shared/form';
