+++ src/widgets/user/NewPassword/ui.tsx
@@ -1,6 +1,6 @@
 import { FC } from 'react';
 
-import { Form } from '@/features/Form';
+import { Form } from '@/entities/Form';
 
 import { Button } from '@/shared/ui/Button';
